gland_back/requirements.txt
uvicorn~=0.23.2
fastapi~=0.103.2
passlib~=1.7.4
pydantic~=2.4.2
SQLAlchemy~=2.0.22
python-jose[cryptography]~=3.3.0
PyMySQL~=1.1.0

pytz~=2023.3.post1
starlette~=0.27.0